% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_connect.R
\name{fleet_connect}
\alias{fleet_connect}
\title{Connect to Pyfleets}
\usage{
fleet_connect(
  api_token = Sys.getenv("DATAFLEETS_API_KEY"),
  coordinator_address = Sys.getenv("DATAFLEETS_COORDINATOR")
)
}
\arguments{
\item{api_token}{API token for configuration.  Default pulls from
environment variables}

\item{coordinator_address}{coordinator address for configuration.
Default pulls from environment variables}
}
\value{
A fleet session object
}
\description{
Connect to Pyfleets
}
\examples{
use_datafleets_conda()
res = fleet_connect()
}
