% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_connect.R
\name{fleet_sql}
\alias{fleet_sql}
\alias{csv_query}
\alias{csv_query_file}
\title{Run SQL Query for Pyfleets}
\usage{
fleet_sql(fleet, query, ...)

csv_query(
  fleet,
  query,
  meta = NULL,
  epsilon = NULL,
  max_contrib = NULL,
  schema = NULL,
  use_cached_result = TRUE
)

csv_query_file(fleet, query, ...)
}
\arguments{
\item{fleet}{An output of \code{\link{fleet_connect}}}

\item{query}{the SQL query to send to pyfleets}

\item{...}{pass through arguments}

\item{meta}{meta data information, passed to \code{QueryParameters}}

\item{epsilon}{epsilon for differential privacy, passed to \code{QueryParameters}}

\item{max_contrib}{maximum contribution, passed to \code{QueryParameters}}

\item{schema}{SQL schema to pass to \code{QueryParameters}}

\item{use_cached_result}{logical passed to \code{QueryParameters}}
}
\value{
A \code{data.frame} of data
}
\description{
Run SQL Query for Pyfleets
}
